refactor(sdk): move preview types and buildPreview to separate preview.ts
Extract FieldMatchMode, PreviewMode, PreviewField, PreviewConfig and
buildPreview from filesystem-serdes.ts into a standalone preview.ts.
filesystem-serdes.ts re-exports them for backward compatibility.
buildPreview is now independently usable without importing filesystem-serdes.
