+++ packages/aws-durable-execution-sdk-js/src/index.ts
@@ -67,6 +67,7 @@ export {
   PreviewField,
   PreviewConfig,
 } from "./utils/serdes/filesystem-serdes";
+export { buildPreview } from "./utils/serdes/preview";
 export { DurableExecutionApiClient } from "./durable-execution-api-client/durable-execution-api-client";
 export {
   createWaitStrategy,
